Moved service file to correct folder
