+++ src/modules/token/fungible/store/service.test.js
