+++ src/redux/actions.js
@@ -7,6 +7,7 @@ export * from '@pos/validator/store/actions/staking';
 export * from '@pos/validator/store/actions/watchList';
 export * from '@update/store/actions/appUpdates';
 export * from 'src/modules/common/store/actions';
-export * from '@token/fungible/utils/service';
+export * from '@token/fungible/store/actions';
+export * from '@token/fungible/store/service';
 export * from '@blockchainApplication/manage/store/action';
 export { default as processLaunchProtocol } from 'src/utils/urlProcessor';
